pyproject.toml
[project]
name = "yt-hindi"
version = "0.1.0"
description = "YouTube to Hindi dubbing pipeline with MCP server"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "yt-dlp>=2024.0.0",
    "mlx-whisper>=0.4.0",
    "google-genai>=1.0.0",
    "edge-tts>=6.1.0",
    "pydub>=0.25.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
]

[project.scripts]
yt-hindi = "yt_hindi.cli:main"
yt-hindi-server = "yt_hindi.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
